requirements.txt
arrow==1.2.3
asgiref==3.6.0
binaryornot==0.4.4
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
cookiecutter==2.2.3
distlib==0.3.6
Django==4.2
django-browser-reload==1.11.0
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-tailwind==3.6.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.12.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
openpyxl==3.1.2
Pillow==9.5.0
pipenv==2023.4.20
platformdirs==3.2.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.22.0
virtualenv-clone==0.5.7
